Consolidate Use/UseTree editor removal into the Removable trait
ast::edit had private inherent ast::Use::remove and
ast::UseTree::remove_with_editor methods that duplicated the
